quick fix: don't fail if we don't have any rust changes

+++ web/internal-scripts/src/rustChange.ts
@@ -20,7 +20,9 @@ const withKnope = async (f: () => Promise<void>) => {
     await f();
   } finally {
     // Restore paths
-    await $`mv ${changesetDir} ${knopeDir}`;
+    if (await fs.exists(changesetDir)) {
+      await $`mv ${changesetDir} ${knopeDir}`;
+    }
     await $`mv ${tsDir} ${changesetDir}`;
   }
 };
